% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arun2010.R
\name{arun2010}
\alias{arun2010}
\title{arun2010}
\usage{
arun2010(dtm, topics = seq(10, 40, by = 10), control = NULL,
  mc.cores = 2L, verbose = FALSE)
}
\arguments{
\item{dtm}{An object of class "\link[tm]{DocumentTermMatrix}" with
term-frequency weighting or an object coercible to a
"\link[slam]{simple_triplet_matrix}" with integer entries.}

\item{topics}{A list with number of topics to compare different models.}

\item{control}{A named list of the control parameters for estimation or an
object of class "\linkS4class{LDAcontrol}".}

\item{mc.cores}{Integer; The number of CPU cores to processes models
simultaneously (using \code{mclapply}).}

\item{verbose}{If false (default), supress all warnings and additional
  information.}
}
\value{
Data-frame with numbers of topics and corresponding values of metric
  (lower is better). Can be directly used by \code{\link{arun2010plot}} to
  draw a plot.
}
\description{
Calculates arun2010 metric for a series of LDA models to estimate the most
preferable number of topics.
}
\section{References}{


  Arun, R., Suresh, V., Veni Madhavan, C. E. & Narasimha Murthy, M. N.
  (2010). On Finding the Natural Number of Topics with Latent Dirichlet
  Allocation: Some Observations. Zaki, M. J., Yu, J. X., Ravindran, B. &
  Pudi, V. (Eds.) In \emph{Advances in Knowledge Discovery and Data Mining},
  Springer Berlin Heidelberg, 6118, 391-402. DOI:
  \href{http://dx.doi.org/10.1073/10.1007/978-3-642-13657-3_43}{10.1007/978-3-642-13657-3_43}.
}
\examples{
library(topicmodels)
data("AssociatedPress", package="topicmodels")
dtm <- AssociatedPress[1:10, ]
arun2010(dtm, topics = 2:10, mc.cores = 2L)
}

