% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arun2010.R
\name{arun2010plot}
\alias{arun2010plot}
\title{arun2010plot}
\usage{
arun2010plot(results)
}
\arguments{
\item{results}{A data.frame with first column "topics" and second "arun2010".}
}
\value{
Returns \code{\link[ggplot2]{ggplot}} object, which can be print or
  used for subsequent tuning by the user desire.
}
\description{
Is a supplementary function to illustrate results of \code{\link{arun2010}}
function.
}
\examples{
library(topicmodels)
data("AssociatedPress", package="topicmodels")
dtm <- AssociatedPress[1:10, ]
metrics <- arun2010(dtm, topics = 2:10, mc.cores = 2L)
arun2010plot(metrics)
}

