% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CaoJuan2009}
\alias{CaoJuan2009}
\title{CaoJuan2009}
\usage{
CaoJuan2009(models)
}
\arguments{
\item{models}{An object of class "\link[topicmodels]{LDA}}
}
\value{
A scalar LDA model score
}
\description{
Implement scoring algorithm
}
