% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Deveaud2014}
\alias{Deveaud2014}
\title{Deveaud2014}
\usage{
Deveaud2014(models)
}
\arguments{
\item{models}{An object of class "\link[topicmodels]{LDA}}
}
\value{
A scalar LDA model score
}
\description{
Implement scoring algorithm
}
