% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/griffiths2004.R
\name{griffiths2004}
\alias{griffiths2004}
\title{griffiths2004}
\usage{
griffiths2004(dtm, topics = seq(10, 40, by = 10), control = NULL,
  mc.cores = 2L, verbose = FALSE)
}
\arguments{
\item{dtm}{An object of class "\link[tm]{DocumentTermMatrix}" with
term-frequency weighting or an object coercible to a
"\link[slam]{simple_triplet_matrix}" with integer entries.}

\item{topics}{A list with number of topics to compare different models.}

\item{control}{A named list of the control parameters for estimation or an
object of class "\linkS4class{LDAcontrol}".}

\item{mc.cores}{Integer; The number of CPU cores to processes models
simultaneously (using \code{mclapply}).}

\item{verbose}{If false (default), supress all warnings and additional
  information.}
}
\value{
Data-frame with numbers of topics and corresponding values of metric
  (higher is better). Can be directly used by \code{\link{griffiths2004plot}}
  to draw a plot.
}
\description{
Calculates griffiths2004 metric for a series of LDA models to estimate the
most preferable number of topics.
}
\section{References}{


  Griffiths, T. L. & Steyvers, M. Finding scientific topics (2004). In
  \emph{Proceedings of the National Academy of Sciences}, National Academy of
  Sciences, 101, 5228-5235. DOI:
  \href{http://dx.doi.org/10.1073/pnas.0307752101}{10.1073/pnas.0307752101}.

  Ponweiser, M. Latent Dirichlet Allocation in R (2012). Diploma Thesis.
  Institute for Statistics and Mathematics, University of Economics and
  Business, Vienna. URL: \url{http://epub.wu.ac.at/id/eprint/3558}.
}
\examples{
library(topicmodels)
data("AssociatedPress", package="topicmodels")
dtm <- AssociatedPress[1:10, ]
griffiths2004(dtm, topics = 2:10, mc.cores = 2L)
}

