% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/griffiths2004.R
\name{griffiths2004plot}
\alias{griffiths2004plot}
\title{griffiths2004plot}
\usage{
griffiths2004plot(results, show.max = TRUE, verbose = FALSE)
}
\arguments{
\item{results}{A data.frame with first column "topics" and second (or many)
contains correspond harmonic mean values (griffiths2004 metrics).}

\item{show.max}{If true (default), find maximum value of metric and point it
on the plot.}

\item{verbose}{If true, produce an additional information.}
}
\value{
Returns \code{\link[ggplot2]{ggplot}} object, which can be print or
  used for subsequent tuning by the user desire.
}
\description{
Is a supplementary function to illustrate results of
\code{\link{griffiths2004}} function.
}
\examples{
library(topicmodels)
data("AssociatedPress", package="topicmodels")
dtm <- AssociatedPress[1:10, ]
metrics <- griffiths2004(dtm, topics = 2:10, mc.cores = 2L)
griffiths2004plot(metrics)
}

