% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldatuning.R
\docType{package}
\name{ldatuning}
\alias{ldatuning}
\title{ldatuning: Tuning of the LDA models parameters}
\description{
A package for identifying the number of topics in a text corpus
by generating LDA models, tuning LDA model parameters, and scoring model
results.
}
