% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Arun2010}
\alias{Arun2010}
\title{Arun2010}
\usage{
Arun2010(models, dtm)
}
\arguments{
\item{models}{An object of class "\link[topicmodels]{LDA}}

\item{dtm}{An object of class "\link[tm]{DocumentTermMatrix}" with
term-frequency weighting or an object coercible to a
"\link[slam]{simple_triplet_matrix}" with integer entries.}
}
\value{
A scalar LDA model score
}
\description{
Implement scoring algorithm
}
