% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Griffiths2004}
\alias{Griffiths2004}
\title{Griffiths2004}
\usage{
Griffiths2004(models, control)
}
\arguments{
\item{models}{An object of class "\link[topicmodels]{LDA}}

\item{control}{A named list of the control parameters for estimation or an
object of class "\linkS4class{LDAcontrol}".}
}
\value{
A scalar LDA model score
}
\description{
Implement scoring algorithm
}
