% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldatuning.R
\docType{package}
\name{ldatuning}
\alias{ldatuning}
\alias{ldatuning-package}
\title{ldatuning: Tuning of the LDA models parameters}
\description{
For this first version only two metrics are implemented:
arun2010, griffiths2004.
}

